+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.6.0
 
 - Add support for additional abilities via a settings value (by undonealex99).
+- Delete an item when reducing the quantity from 1 to 0.
 
 # 2.5.4